+++ src/main/java/com/telepathicgrunt/tact/BlockModifications.java
@@ -17,7 +17,7 @@ import java.util.function.Supplier;
 
 public class BlockModifications {
 
-    static void setupModifications(final FMLCommonSetupEvent event) {
+    static void doModifications(final FMLCommonSetupEvent event) {
         if (Config.REPLACEABLE_SMALL_PLANTS.get()) {
             makeReplaceable(new ResourceLocation(TACT.ALEXS_CAVES_MODID, "underweed"));
             makeReplaceable(new ResourceLocation(TACT.ALEXS_CAVES_MODID, "tree_star"));