+++ src/main/java/com/telepathicgrunt/tact/mixin/ItemAccessor.java
@@ -1,0 +1,15 @@
+package com.telepathicgrunt.tact.mixin;
+
+import net.minecraft.world.item.Item;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.UUID;
+
+@Mixin(Item.class)
+public interface ItemAccessor {
+    @Accessor
+    static UUID getBASE_ATTACK_DAMAGE_UUID() {
+        throw new UnsupportedOperationException();
+    }
+}