+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 #### Entities:
 Added 6 new configs to control the health recovery amount and health recovery speed of Tremorzilla.
+
+#### Items:
+Fixed Dreadbow Arrow config not actually applying.