+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### **(V.1.2.0 Changes) (1.20.1 Minecraft)**
+### **(V.1.2.1 Changes) (1.20.1 Minecraft)**
 
-#### Major:
-Now requires Alex's Caves v2.0.0! Cavity Caves are now tagged properly and their mobs have health, armor, and attack configs!
+#### Configs:
+Fixed Corrodent using Boundroid armor config value.
+
+Fixed Sweetish Fish using Subterrandon health config value.